+++ src/assets.ts
@@ -65,6 +65,7 @@ export function assetToName(asset: Asset): string | null {
   if (asset == Asset.STRK) return "STRK";
   if (asset == Asset.WIF) return "WIF";
   if (asset == Asset.RNDR) return "RNDR";
+  if (asset == Asset.TNSR) return "TNSR";
   if (asset == null) return null; // Some things, like clock callbacks, are for all assets and return asset=null
   return "UNDEFINED";
 }
@@ -86,6 +87,7 @@ export function nameToAsset(name: string): Asset {
   if (name == "STRK") return Asset.STRK;
   if (name == "WIF") return Asset.WIF;
   if (name == "RNDR") return Asset.RNDR;
+  if (name == "TNSR") return Asset.TNSR;
   return Asset.UNDEFINED;
 }
 
@@ -110,6 +112,7 @@ export function toProgramAsset(asset: Asset): any {
   if (asset == Asset.STRK) return { strk: {} };
   if (asset == Asset.WIF) return { wif: {} };
   if (asset == Asset.RNDR) return { rndr: {} };
+  if (asset == Asset.TNSR) return { tnsr: {} };
   return { undefined: {} };
 }
 
@@ -161,6 +164,9 @@ export function fromProgramAsset(asset: any): Asset {
   }
   if (objectEquals(asset, { rndr: {} })) {
     return Asset.RNDR;
+  }
+  if (objectEquals(asset, { tnsr: {} })) {
+    return Asset.TNSR;
   }
   return Asset.UNDEFINED;
 }
@@ -215,6 +221,9 @@ export function assetToIndex(asset: Asset): number {
     case Asset.RNDR: {
       return 15;
     }
+    case Asset.TNSR: {
+      return 16;
+    }
   }
   return 255; // Undefined is 255 onchain
 }
@@ -268,6 +277,9 @@ export function indexToAsset(index: number): Asset {
     }
     case 15: {
       return Asset.RNDR;
+    }
+    case 16: {
+      return Asset.TNSR;
     }
   }
   return Asset.UNDEFINED;