+++ src/types/zeta.ts
@@ -6271,6 +6271,142 @@ export type Zeta = {
       ]
     },
     {
+      "name": "adminSetOrderState",
+      "accounts": [
+        {
+          "name": "authority",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "cancelAccounts",
+          "accounts": [
+            {
+              "name": "state",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "marginAccount",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "dexProgram",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "serumAuthority",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "openOrders",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "market",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "bids",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "asks",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "eventQueue",
+              "isMut": true,
+              "isSigner": false
+            }
+          ]
+        }
+      ],
+      "args": [
+        {
+          "name": "asset",
+          "type": {
+            "defined": "Asset"
+          }
+        }
+      ]
+    },
+    {
+      "name": "adminForceCancelOrders",
+      "accounts": [
+        {
+          "name": "authority",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "cancelAccounts",
+          "accounts": [
+            {
+              "name": "state",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "marginAccount",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "dexProgram",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "serumAuthority",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "openOrders",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "market",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "bids",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "asks",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "eventQueue",
+              "isMut": true,
+              "isSigner": false
+            }
+          ]
+        }
+      ],
+      "args": [
+        {
+          "name": "asset",
+          "type": {
+            "defined": "Asset"
+          }
+        }
+      ]
+    },
+    {
       "name": "forceCancelOrdersV2",
       "accounts": [
         {
@@ -6438,6 +6574,52 @@ export type Zeta = {
           "type": {
             "defined": "Asset"
           }
+        }
+      ]
+    },
+    {
+      "name": "adminCrankEventQueue",
+      "accounts": [
+        {
+          "name": "state",
+          "isMut": false,
+          "isSigner": false
+        },
+        {
+          "name": "authority",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "market",
+          "isMut": true,
+          "isSigner": false
+        },
+        {
+          "name": "eventQueue",
+          "isMut": true,
+          "isSigner": false
+        },
+        {
+          "name": "dexProgram",
+          "isMut": false,
+          "isSigner": false
+        },
+        {
+          "name": "serumAuthority",
+          "isMut": false,
+          "isSigner": false
+        },
+        {
+          "name": "openOrders",
+          "isMut": true,
+          "isSigner": false
+        }
+      ],
+      "args": [
+        {
+          "name": "eventsToCrank",
+          "type": "u16"
         }
       ]
     },
@@ -7077,7 +7259,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -7086,7 +7268,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u64",
-                9
+                8
               ]
             }
           },
@@ -7095,7 +7277,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -7104,7 +7286,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u64",
-                9
+                8
               ]
             }
           },
@@ -7115,7 +7297,7 @@ export type Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                16
+                17
               ]
             }
           },
@@ -7126,7 +7308,7 @@ export type Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                9
+                8
               ]
             }
           },
@@ -7137,7 +7319,7 @@ export type Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                16
+                17
               ]
             }
           },
@@ -7148,7 +7330,7 @@ export type Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                9
+                8
               ]
             }
           },
@@ -7157,7 +7339,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -7166,7 +7348,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u64",
-                9
+                8
               ]
             }
           },
@@ -7175,7 +7357,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -7184,7 +7366,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -7193,7 +7375,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -7202,7 +7384,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -7211,7 +7393,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -7220,7 +7402,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -7229,7 +7411,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -7238,7 +7420,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -7249,7 +7431,7 @@ export type Zeta = {
                 {
                   "defined": "PerpParameters"
                 },
-                16
+                17
               ]
             }
           },
@@ -7260,7 +7442,7 @@ export type Zeta = {
                 {
                   "defined": "PerpParameters"
                 },
-                9
+                8
               ]
             }
           },
@@ -7271,7 +7453,7 @@ export type Zeta = {
                 {
                   "defined": "MarginParameters"
                 },
-                16
+                17
               ]
             }
           },
@@ -7282,7 +7464,7 @@ export type Zeta = {
                 {
                   "defined": "MarginParameters"
                 },
-                9
+                8
               ]
             }
           },
@@ -7293,7 +7475,7 @@ export type Zeta = {
                 {
                   "defined": "Product"
                 },
-                16
+                17
               ]
             }
           },
@@ -7304,7 +7486,7 @@ export type Zeta = {
                 {
                   "defined": "Product"
                 },
-                9
+                8
               ]
             }
           },
@@ -7313,7 +7495,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -7322,7 +7504,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -7343,7 +7525,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "bool",
-                16
+                17
               ]
             }
           },
@@ -7352,7 +7534,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "bool",
-                9
+                8
               ]
             }
           },
@@ -7744,7 +7926,7 @@ export type Zeta = {
                 {
                   "defined": "HaltStateV2"
                 },
-                16
+                17
               ]
             }
           },
@@ -7755,7 +7937,7 @@ export type Zeta = {
                 {
                   "defined": "HaltStateV2"
                 },
-                9
+                8
               ]
             }
           },
@@ -7768,7 +7950,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u32",
-                16
+                17
               ]
             }
           },
@@ -7777,7 +7959,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u32",
-                9
+                8
               ]
             }
           },
@@ -7786,7 +7968,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u32",
-                16
+                17
               ]
             }
           },
@@ -7795,7 +7977,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u32",
-                9
+                8
               ]
             }
           },
@@ -8195,11 +8377,15 @@ export type Zeta = {
             "type": "u8"
           },
           {
+            "name": "referredTimestamp",
+            "type": "u64"
+          },
+          {
             "name": "padding",
             "type": {
               "array": [
                 "u8",
-                22
+                14
               ]
             }
           }
@@ -8246,7 +8432,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u8",
-                16
+                17
               ]
             }
           },
@@ -8255,7 +8441,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u8",
-                9
+                8
               ]
             }
           },
@@ -8270,7 +8456,7 @@ export type Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                16
+                17
               ]
             }
           },
@@ -8281,7 +8467,7 @@ export type Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                9
+                8
               ]
             }
           },
@@ -8292,7 +8478,7 @@ export type Zeta = {
                 {
                   "defined": "ProductLedger"
                 },
-                16
+                17
               ]
             }
           },
@@ -8303,7 +8489,7 @@ export type Zeta = {
                 {
                   "defined": "ProductLedger"
                 },
-                9
+                8
               ]
             }
           },
@@ -8320,7 +8506,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -8329,7 +8515,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u64",
-                9
+                8
               ]
             }
           },
@@ -9113,7 +9299,7 @@ export type Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -9852,6 +10038,9 @@ export type Zeta = {
           },
           {
             "name": "RNDR"
+          },
+          {
+            "name": "TNSR"
           },
           {
             "name": "UNDEFINED"
@@ -11395,6 +11584,16 @@ export type Zeta = {
       "code": 6171,
       "name": "AirdropCommunityAlreadySet",
       "msg": "Airdrop community already set, cannot set again"
+    },
+    {
+      "code": 6172,
+      "name": "RemainingOrdersInBookAfterPrune",
+      "msg": "User still has orders in orderbook after prune"
+    },
+    {
+      "code": 6173,
+      "name": "OODoesntMatch",
+      "msg": "Can only pop events for open orders passed in"
     }
   ]
 };
@@ -17672,6 +17871,142 @@ export const IDL: Zeta = {
       ]
     },
     {
+      "name": "adminSetOrderState",
+      "accounts": [
+        {
+          "name": "authority",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "cancelAccounts",
+          "accounts": [
+            {
+              "name": "state",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "marginAccount",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "dexProgram",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "serumAuthority",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "openOrders",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "market",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "bids",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "asks",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "eventQueue",
+              "isMut": true,
+              "isSigner": false
+            }
+          ]
+        }
+      ],
+      "args": [
+        {
+          "name": "asset",
+          "type": {
+            "defined": "Asset"
+          }
+        }
+      ]
+    },
+    {
+      "name": "adminForceCancelOrders",
+      "accounts": [
+        {
+          "name": "authority",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "cancelAccounts",
+          "accounts": [
+            {
+              "name": "state",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "marginAccount",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "dexProgram",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "serumAuthority",
+              "isMut": false,
+              "isSigner": false
+            },
+            {
+              "name": "openOrders",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "market",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "bids",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "asks",
+              "isMut": true,
+              "isSigner": false
+            },
+            {
+              "name": "eventQueue",
+              "isMut": true,
+              "isSigner": false
+            }
+          ]
+        }
+      ],
+      "args": [
+        {
+          "name": "asset",
+          "type": {
+            "defined": "Asset"
+          }
+        }
+      ]
+    },
+    {
       "name": "forceCancelOrdersV2",
       "accounts": [
         {
@@ -17839,6 +18174,52 @@ export const IDL: Zeta = {
           "type": {
             "defined": "Asset"
           }
+        }
+      ]
+    },
+    {
+      "name": "adminCrankEventQueue",
+      "accounts": [
+        {
+          "name": "state",
+          "isMut": false,
+          "isSigner": false
+        },
+        {
+          "name": "authority",
+          "isMut": false,
+          "isSigner": true
+        },
+        {
+          "name": "market",
+          "isMut": true,
+          "isSigner": false
+        },
+        {
+          "name": "eventQueue",
+          "isMut": true,
+          "isSigner": false
+        },
+        {
+          "name": "dexProgram",
+          "isMut": false,
+          "isSigner": false
+        },
+        {
+          "name": "serumAuthority",
+          "isMut": false,
+          "isSigner": false
+        },
+        {
+          "name": "openOrders",
+          "isMut": true,
+          "isSigner": false
+        }
+      ],
+      "args": [
+        {
+          "name": "eventsToCrank",
+          "type": "u16"
         }
       ]
     },
@@ -18478,7 +18859,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -18487,7 +18868,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u64",
-                9
+                8
               ]
             }
           },
@@ -18496,7 +18877,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -18505,7 +18886,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u64",
-                9
+                8
               ]
             }
           },
@@ -18516,7 +18897,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                16
+                17
               ]
             }
           },
@@ -18527,7 +18908,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                9
+                8
               ]
             }
           },
@@ -18538,7 +18919,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                16
+                17
               ]
             }
           },
@@ -18549,7 +18930,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                9
+                8
               ]
             }
           },
@@ -18558,7 +18939,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -18567,7 +18948,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u64",
-                9
+                8
               ]
             }
           },
@@ -18576,7 +18957,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -18585,7 +18966,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -18594,7 +18975,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -18603,7 +18984,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -18612,7 +18993,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -18621,7 +19002,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -18630,7 +19011,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -18639,7 +19020,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -18650,7 +19031,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "PerpParameters"
                 },
-                16
+                17
               ]
             }
           },
@@ -18661,7 +19042,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "PerpParameters"
                 },
-                9
+                8
               ]
             }
           },
@@ -18672,7 +19053,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "MarginParameters"
                 },
-                16
+                17
               ]
             }
           },
@@ -18683,7 +19064,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "MarginParameters"
                 },
-                9
+                8
               ]
             }
           },
@@ -18694,7 +19075,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "Product"
                 },
-                16
+                17
               ]
             }
           },
@@ -18705,7 +19086,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "Product"
                 },
-                9
+                8
               ]
             }
           },
@@ -18714,7 +19095,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                16
+                17
               ]
             }
           },
@@ -18723,7 +19104,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "publicKey",
-                9
+                8
               ]
             }
           },
@@ -18744,7 +19125,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "bool",
-                16
+                17
               ]
             }
           },
@@ -18753,7 +19134,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "bool",
-                9
+                8
               ]
             }
           },
@@ -19145,7 +19526,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "HaltStateV2"
                 },
-                16
+                17
               ]
             }
           },
@@ -19156,7 +19537,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "HaltStateV2"
                 },
-                9
+                8
               ]
             }
           },
@@ -19169,7 +19550,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u32",
-                16
+                17
               ]
             }
           },
@@ -19178,7 +19559,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u32",
-                9
+                8
               ]
             }
           },
@@ -19187,7 +19568,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u32",
-                16
+                17
               ]
             }
           },
@@ -19196,7 +19577,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u32",
-                9
+                8
               ]
             }
           },
@@ -19596,11 +19977,15 @@ export const IDL: Zeta = {
             "type": "u8"
           },
           {
+            "name": "referredTimestamp",
+            "type": "u64"
+          },
+          {
             "name": "padding",
             "type": {
               "array": [
                 "u8",
-                22
+                14
               ]
             }
           }
@@ -19647,7 +20032,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u8",
-                16
+                17
               ]
             }
           },
@@ -19656,7 +20041,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u8",
-                9
+                8
               ]
             }
           },
@@ -19671,7 +20056,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                16
+                17
               ]
             }
           },
@@ -19682,7 +20067,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "AnchorDecimal"
                 },
-                9
+                8
               ]
             }
           },
@@ -19693,7 +20078,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "ProductLedger"
                 },
-                16
+                17
               ]
             }
           },
@@ -19704,7 +20089,7 @@ export const IDL: Zeta = {
                 {
                   "defined": "ProductLedger"
                 },
-                9
+                8
               ]
             }
           },
@@ -19721,7 +20106,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -19730,7 +20115,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u64",
-                9
+                8
               ]
             }
           },
@@ -20514,7 +20899,7 @@ export const IDL: Zeta = {
             "type": {
               "array": [
                 "u64",
-                16
+                17
               ]
             }
           },
@@ -21253,6 +21638,9 @@ export const IDL: Zeta = {
           },
           {
             "name": "RNDR"
+          },
+          {
+            "name": "TNSR"
           },
           {
             "name": "UNDEFINED"
@@ -22796,6 +23184,16 @@ export const IDL: Zeta = {
       "code": 6171,
       "name": "AirdropCommunityAlreadySet",
       "msg": "Airdrop community already set, cannot set again"
+    },
+    {
+      "code": 6172,
+      "name": "RemainingOrdersInBookAfterPrune",
+      "msg": "User still has orders in orderbook after prune"
+    },
+    {
+      "code": 6173,
+      "name": "OODoesntMatch",
+      "msg": "Can only pop events for open orders passed in"
     }
   ]
 };